Update Convert-VMtoManagedDisks.ps1
Changelog
0.5 - The "I remembered I need to push a commit" update
Added Versioning Tag
Updated comments for what is and isn't implemented
Changed OutReport flag to accept no input (And default properly to Documents)
Updated multiple $null checks to better  practice per https://learn.microsoft.com/en-us/powershell/utility-modules/psscriptanalyzer/rules/possibleincorrectcomparisonwithnull?view=ps-modules
Removed Starttime - Endtime speed test code
Added continuePrompt function to replace multiple continue loops
Moved VM Rebuild code to be a function
Added validation before Starting Disk Migration and set to run after successful VM Creation
A LOT of misc code flow changes
ReRemembered how to update Github
Changed Revert-VM internal function to revertVM for better readability

+++ Convert-VMtoManagedDisks.ps1
@@ -1,5 +1,4 @@
-﻿
-    <#
+﻿    <#
     .SYNOPSIS
         Converts all disks on a Azure Stack Hub Virtual machine from Unmanaged Disks to Managed Disks and automates VM deletion and recreation
 
@@ -8,6 +7,7 @@
 
     .PARAMETER ADFS
         Indicates the Stamp is an ADFS authentication based stamp. Exclusive from flag -AADTenantName
+        CURRENTLY NOT IMPLEMENTED
 
     .PARAMETER FQDN
         Fullly qualified domain name for a Stack Hub stamp
@@ -19,11 +19,12 @@
         The VM's name in the Azure Stack Portal
 
     .EXAMPLE
-        PS> Convert-VMtoManagedDisks.ps1
+        PS> Convert-VMtoManagedDisks.ps1 -AADTenantName Contosolabs.onmicrosoft.com -FQDN USStamp1.Contoso.com -OutReport C:\Temp\AzSVM1Legacy\ -VMName AzSVM1Legacy
 
     .EXAMPLE
         PS> Get-AZSSupportVM
-
+    .Version
+        0.1
     #>
 
     [CmdletBinding(DefaultParameterSetName="Default")]
@@ -31,23 +32,62 @@ param(
 
      
     [Parameter(Mandatory=$True, ValueFromPipeline=$false, ParameterSetName="AAD")]
-    $AADTenantName,
+    [string]$AADTenantName,
 
-    [Parameter(Mandatory=$True, ValueFromPipeline=$false, ParameterSetName="ADFS")]
+    [Parameter(Mandatory=$True, ValueFromPipeline=$false, ParameterSetName="ADFS")]#CURRENTLY NOT IMPLMENTED
     [Switch]$ADFS,
 
     [Parameter(Mandatory=$True, ValueFromPipeline=$false)]
     [string]$FQDN,
 
-    [Parameter(Mandatory=$True, ValueFromPipeline=$false)]
+    [Parameter(Mandatory=$false, ValueFromPipeline=$false)]
     [ValidateScript({Test-Path $_ -PathType Container})]
-    $OutReport = "$env:USERPROFILE\Documents",
+    [string]$OutReport = "$env:USERPROFILE\Documents",
 
     [Parameter(Mandatory=$True, ValueFromPipeline=$false)]
-    $VMName
+    [string]$VMName
         
 
 )
+
+#####Define Internal Functions
+#Loop Continue Function
+function continuePrompt{
+param(
+    [Parameter(Mandatory=$True, ValueFromPipeline=$false)]
+    [string]$Prompt,
+    [Parameter(Mandatory=$false, ValueFromPipeline=$false)]
+    [string]$ReturnNo
+      )
+Do{
+$Continue = Read-Host -Prompt $Prompt # "Continue Migration Anyway (Y/N)?"
+switch -wildcard ($Continue){
+N{Write-Host "Stopping" ; $Go = $true ; return $ReturnNo}
+Y{Write-Host "Continuing" ; $Go = $true ; continue}
+default {Write-Host "Invalid Response, please enter Y or N"
+          $Go = $false
+$ErrorNoise= @(554.37,659.26,622.25,659.26,415.3)
+$ErrorNoise | %{[console]::Beep($_,$(125))}
+
+        }
+                                   }
+                                                        
+  }Until($Go -eq $true)
+                            }
+                         
+
+
+#VM Recovery/Rebuild function in case errors occur later on
+function revertVM{
+Write-Host "Attempting to recover VM"
+$VMRebuild = $azsVM
+$VMRebuild.StorageProfile.OsDisk.CreateOption = "Attach"
+$VMRebuild.StorageProfile.DataDisks | %{$_.CreateOption = "Attach"}
+$VMRebuild.StorageProfile.ImageReference = $null
+$VMRebuild.OSProfile = $null
+New-AzVM -VM $VMRebuild -ResourceGroupName $azsVM.ResourceGroupName -Location $azsVM.Location
+return                                   
+                    }
 
 #Check Prereqs are installed
 try{Import-Module az -ErrorAction Stop}
@@ -60,28 +100,19 @@ Catch {
         Write-Host -ForegroundColor Red -BackgroundColor Black $Error[0]
         exit
         }
-if ($(Get-Module -Name az) -notlike $null ){write-host "Azure az module loaded successfully"}
+if ($null -ne $(Get-Module -Name az)  ){write-host "Azure az module loaded successfully"}
 
-##Worth checking if AzureRM is installed?
+#Verify if AzureRM module is installed. If so ask the user if they want to continue (Not forcing script stop in case AzureRM is installed in a configuration where it wouldn't interact with this PoSH session (IE, Az installed in 5.2 and AzureRM in 7)
 
-if($(get-module -ListAvailable azurerm*) -notlike $null){
-Do{
+if($null -ne $(get-module -ListAvailable azurerm*) ){
 Write-Warning "AzureRM Module detected as installed on system, issues may occur if module is used alongside az module. Advise removing module before continuning"
-$AzureRMContinue = Read-Host -Prompt "Continue Migration Anyway (Y/N)?"
-switch -wildcard ($AzureRMContinue){
-N*{write "Stopping" ; $Go = $true ; exit}
-Y*{write "Continuing" ; $Go = $true ; continue}
-default {"Invalid Response"
-          $Go = $false
-          [console]::Beep(1000,300)  
-        }
-                                   }
-                                                        
-  }Until($Go -eq $true)#Check if an AzureStackUser enviornment exists and prompt user if it doesn't match what we'd want to use
+continuePrompt -Prompt "Continue Migration Anyway (Y/N)?"
+  
+  #Login to Azure Stack from an  AzureStackUser enviornment exists and prompt user if it doesn't match what we'd want to use
   }
 #! Check Azure Stack Connects successfully
 Try {
-Write-Host -BackgroundColor Gray -ForegroundColor Green "Connecting to Azure Stack..."
+Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Connecting to Azure Stack..."
 Add-AzEnvironment –Name ‘AzureStackUser’ -ArmEndpoint "https://management.$FQDN" | Out-Null
 # Set your AAD tenant name
 $AuthEndpoint = (Get-AzEnvironment -Name "AzureStackUser").ActiveDirectoryAuthority.TrimEnd('/')
@@ -95,39 +126,12 @@ Write-Warning "Cannot connect to Azure Stack environment. Please check your cred
 Break
 }
 
-# Define recovery function in case errors occur later on
-function Revert-VM{do{$Recover = Read-Host -Prompt "Attempt to recover VM? (Y/N)"
-                    switch -Wildcard ($Recover){
-                                    #Show data lost and so VM can be manually triaged
-                                    N*{write-host "Ending Script, Note both managed and unmanaged disks for the following still exist:"
-                                    $DisksCreated
-
-                                    Write-host "
-                                                The Following Disks were not Created"
-                                                $DisksMissed
-                                    }
-                                    #Attempt to Recover VM
-                                    Y*{            Write-Host "Attempting to recover VM"
-                                        $VMRebuild = $azsVM
-                                        $VMRebuild.StorageProfile.OsDisk.CreateOption = "Attach"
-                                        $VMRebuild.StorageProfile.DataDisks | %{$_.CreateOption = "Attach"}
-                                        $VMRebuild.StorageProfile.ImageReference = $null
-                                        $VMRebuild.OSProfile = $null
-                                        New-AzVM -VM $VMRebuild -ResourceGroupName $azsVM.ResourceGroupName -Location $azsVM.Location
-                                    return
-                                    }
-                                    #Default
-                                    default {"Invalid Response"}
-                                    }
-        }until ($Recover -like "Y*" -or $Recover -like "N*")
-
-                    }
 # Get all Azure Stack Tenant Subscriptions and look for the VM that you want to convert
 # The source and target VM will be created
-Write-Host -BackgroundColor Gray -ForegroundColor Green "Get all Azure Stack Tenant Subscriptions..."
+Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Getting all Azure Stack Tenant Subscriptions..."
 $azsSubs = Get-AzSubscription
 
-Write-Host -BackgroundColor Gray -ForegroundColor Green "Locate Azure Stack VM ($VMName)..."
+Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Locate Azure Stack VM ($VMName)..."
 $azsVMs = $azsSubs | ForEach-Object {Select-AzSubscription $_ | Out-Null; Get-AzVM | Where-Object {$_.Name -eq $VMName -and $_.StorageProfile.OsDisk.ManagedDisk -like $null} }
 
 
@@ -191,7 +195,7 @@ $azsVMstoragediag = $azsVMstoragediag.Split('.')[0]
 #####Availability Set#####
 #Get Availability Set to add new VM to
 If ($azsVM.AvailabilitySetReference.Id -ne $null) {
-Write-Host -BackgroundColor Gray -ForegroundColor Green "Check if the VM that you want to convert is a member of an AvailabilitySet..."
+Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Check if the VM that you want to convert is a member of an AvailabilitySet..."
 $avSetName = $azsVM.AvailabilitySetReference.Id.Split('/')[-1]
 $avSet = Get-AzAvailabilitySet -ResourceGroupName $azsVM.ResourceGroupName -Name $avSetName
 }
@@ -376,7 +380,7 @@ p.List table td:first-child {
 
  <tr >
   <td colspan=4 valign=top >
-    <p class=Grid>" + $($avSetReport | ConvertTo-Html -Fragment ).replace("&lt;b&gt;","<b>").replace("&lt;/b&gt;","</b>") + "</p>
+    <p class=grid>" + $($avSetReport | ConvertTo-Html -Fragment ).replace("&lt;b&gt;","<b>").replace("&lt;/b&gt;","</b>") + "</p>
   </td>
  </tr>
  <tr>
@@ -406,16 +410,10 @@ start migration from Unmanaged to Managed Disks this process must be completed i
 Would you like to continue to the next phase of Migration?:")
 
 #####LAST CHANCE FOR CX TO STOP PROCESS IN CASE OF ERROR#####
-$Continue = Read-Host -Prompt "Continue Migration (Y/N)"
-switch -wildcard ($Continue){
-N*{write "Stopping" ; return}
-Y*{write "Continuing" ; continue}
-default {"Invalid Response"
-            }
-}
+$MigrationPrecheck = continuePrompt -Prompt "Continue Migration (Y/N)" -ReturnNo "EndMigration"
+if($MigrationPrecheck -eq "EndMigration"){write-host "Exiting Script" ; exit}
 
-$Starttime = Get-Date
-# Remove the VM, keeping the disks 
+# Remove the VM config, keeping the disks 
 Write-Output -InputObject "Removing old virtual machine"
 Remove-AzVM -Name $VMName -ResourceGroupName $azsVM.ResourceGroupName -Force
 
@@ -441,11 +439,11 @@ $DiskConfig = New-AzDiskConfig -AccountType $azsStorage.Sku.Name -Location $azsV
 if ($azsdisk.OsType -ne $null) { 
    
     # Create one of 2 possible VM configs, adding tags here 
-    Write-Host -BackgroundColor Gray -ForegroundColor Green "Creating virtual machine configuration..."
+    Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Creating virtual machine configuration..."
     #Make sure to add VM to the Availability Set if it's already part of one
 
     If ($avSetName) {
-    Write-Host -BackgroundColor Gray -ForegroundColor Green "Configuring the VM $($azsVM.Name) to be added to managed availability set ($avSetName)"
+    Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Configuring the VM $($azsVM.Name) to be added to managed availability set ($avSetName)"
     $VirtualMachine = New-AzVMConfig -VMName $azsVM.Name -VMSize $azsVM.HardwareProfile.VmSize -Tags $azTags -AvailabilitySetId $avSet.Id 
     }
 
@@ -457,7 +455,7 @@ if ($azsdisk.OsType -ne $null) {
     # Create OS managed disk.
     #TODO: ADD PROPER ERROR HANDLING IN CASE DISK DOESN'T GET CREATED
     #TODO: Specific Error for duplicate disk ErrorMessage: Changing property 'sourceUri' is not allowed for existing disk 'Diskname'. Change Disk name to include a GUID and output name
-    Write-Host -BackgroundColor Gray -ForegroundColor Green "Creating OS managed disk with name: $($azsVM.Name)-$($azsdisk.Name)"
+    Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Creating OS managed disk with name: $($azsVM.Name)-$($azsdisk.Name)"
     $OsDisk = New-AzDisk -DiskName "$($azsVM.Name)-$($azsdisk.Name)" -Disk $DiskConfig -ResourceGroupName $azsVM.ResourceGroupName
     if(!$OsDisk){Write-Error "OS DISK CREATION FAILED, EXITING"
                 return}
@@ -475,7 +473,7 @@ if ($azsdisk.OsType -ne $null) {
 }
 Else {
     # Create Data managed disks.
-    Write-Host -BackgroundColor Gray -ForegroundColor Green "Create the Data managed disk: $($azsdisk.Name)"
+    Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Create the Data managed disk: $($azsdisk.Name)"
     #Clean up last variable
     $DataDisk = $null 
     #Create new Disk
@@ -499,19 +497,20 @@ $DisksMissed | select -Property @{Name="Lun";Expression={if($_.Lun -notlike $nul
                                 Caching,`
                                 @{Name="Disk Location";Expression={$_.vhd.uri}}
 Write-Warning "THE FOLLOWING DISKS ABOVE COULD NOT BE CREATED"
-do{
-$ContinueDiskDeletion = Read-Host -Prompt "Continue Migration (Y/N)"
-switch -wildcard ($Continue){
-N*{write "Stopping"
-Revert-VM
- return}
-Y*{write "Continuing"
-   continue}
-default {"Invalid Response"}
+$DiskCountFail = continuePrompt -Prompt "Continue Migration (Y/N)?" -ReturnNo "RebuildVM"
+if($DiskCountFail -eq "RebuildVM"){
+    Write-Warning "ATTEMPT REBUILD OF UNMANAGED VM?"
+    $TriggerRebuild = continuePrompt -Prompt "Rebuild VM (Y/N)?" -ReturnNo "CancelRebuild"
+if($TriggerRebuild -eq "CancelRebuilld"){
+                                        Write-host "The Following Disks were not Created"
+                                        $DisksMissed  
+                               }
 
+if($null -eq $TriggerRebuild){revertVM}
+write-host "Ending Script, Note both managed and unmanaged disks for the following still exist:"
+                                    $DisksCreated
+exit
 }
-}until($ContinueDiskDeletion -like "Y*" -or $ContinueDiskDeletion -like "N*")
-
 }
 
 #Add NICs to config
@@ -528,40 +527,43 @@ $VirtualMachine = Set-AzVMBootDiagnostic -VM $VirtualMachine -Enable -StorageAcc
 
 #Convert Availability set (IF needed)
 if($avSet -ne $null){
-Write-Host -BackgroundColor Gray -ForegroundColor Green "Convert the availability set ($avSetName) to a managed availability set..."
+Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Convert the availability set ($avSetName) to a managed availability set..."
 Update-AzAvailabilitySet -AvailabilitySet $avSet -Sku Aligned | Out-Null
                     }
 
 
 # Create the new virtual machine with managed disks and boot
 ##TODO Output VM with get-vm to show to the user it's complete
-Write-Host -BackgroundColor Gray -ForegroundColor Green "Create the new virtual machine ($($azsVM.Name)) with managed disks..."
+Write-Host -BackgroundColor DarkGray -ForegroundColor Green "Create the new virtual machine ($($azsVM.Name)) with managed disks..."
 try{New-AzVM -VM $VirtualMachine -ResourceGroupName $azsVM.ResourceGroupName -Location $azsVM.Location
 }catch{Write-Warning "Unable to create vm for reason" $Error[0] 
         return}
 
 
-##TODO wrap this so it adds a warn prompt before starting this process
-Write-Warning "Cleaning up old disks"
-
-foreach ($azsdisk in $DisksCreated) { 
-#Get VM Storage account location for Unmanaged Disk
-$storageacc = ([System.Uri]$azsdisk.Vhd.Uri).Host.Split('.')[0]
-$azsStorage = Get-AzStorageAccount -ResourceGroupName $azsVM.ResourceGroupName -Name $storageacc
-Set-AzCurrentStorageAccount -Name $storageacc -ResourceGroupName $azsVM.ResourceGroupName | Out-Null
-$storageContainerName = (Get-AzStorageContainer).Name
-Write-Warning "Do you want to delete the following"
-$azdisk| select -Property @{Name="Lun";Expression={if($_.Lun -notlike $null){$_.Lun} else{"OS Disk for "+$_.OSType}}},`
-                                Name,`
-                                DiskSizeGB,`
-                                Caching,`
-                                @{Name="Disk Location";Expression={$_.vhd.uri}}
-Remove-AzStorageBlob -Container $storageContainerName -Blob "$($azsdisk.Vhd.Uri.Split('/')[-1])" -Force -Confirm:$True
-}
-
-$Endtime = Get-Date
-
-$Allotted = $Endtime - $Starttime
+Write-Warning "Migration complete, next phase involves removing old managed disks Note this is A DESTRUCTIVE ACTION"
+$DiskCleanup = continuePrompt -Prompt "Begin disk cleanup? (Y/N):" -ReturnNo "SkipCleanup"
+if($null -eq $DiskCleanup){
+    foreach ($azsdisk in $DisksCreated) { 
+        Write-Host ""
+        #Get VM Storage account location for Unmanaged Disk
+        $storageacc = ([System.Uri]$azsdisk.Vhd.Uri).Host.Split('.')[0]
+        $azsStorage = Get-AzStorageAccount -ResourceGroupName $azsVM.ResourceGroupName -Name $storageacc
+        Set-AzCurrentStorageAccount -Name $storageacc -ResourceGroupName $azsVM.ResourceGroupName | Out-Null
+        $storageContainerName = (Get-AzStorageContainer).Name
+        Write-Warning "Starting removal for the below Disk. Do you want to delete the following"
+        $azsdisk| select -Property @{Name="Lun";Expression={if($_.Lun -notlike $null){$_.Lun} else{"OS Disk for "+$_.OSType}}},`
+                                        Name,`
+                                        DiskSizeGB,`
+                                        Caching,`
+                                        @{Name="Disk Location";Expression={$_.vhd.uri}}
+        $DiskConfirm = $null
+        $DiskConfirm = continuePrompt -Prompt "Delete (Y/N)"
+        if($null -eq $DiskConfirm){Remove-AzStorageBlob -Container $storageContainerName -Blob "$($azsdisk.Vhd.Uri.Split('/')[-1])" -Force -Confirm:$True}
+                                        }
+                        }
+if($DiskCleanup -eq "SkipCleanup"){Write-Warning "Disks have been left undeleted. Advise cleanup at a later date"
+$DisksCreated}
+##TODO Figure out how to PROPERLY move VM Extensions with arbitrary data
 #Add Extensions and verify as added
 #New-AzConnectedMachineExtension
